refactor: Home 페이지 컴포넌트 리팩토링 메인 컨텐츠 생성 로직 분리
