+++ src/Pages/index.ts
@@ -1,3 +1,3 @@
-import Home from './Home';
+import Home from './Home/Home';
 
 export { Home };
